+++ CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to HaloCommunityBot will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.29] - 2026-05-14
+
+### Changed
+
+* Fix startup crash from slash command CancellationToken parameter
+
+## [1.1.28] - 2026-05-14
+
+### Changed
+
+* Add token env fallback and startup diagnostics
+
 ## [1.1.27] - 2026-05-14
 
 ### Changed